loader -> vue
